pyproject.toml
[project]
name = "cmdai"
version = "1.0"
description = "A familiar and easy to use terminal interface for OpenAI and/or Ollama"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "textual>=0.47.0",
    "rich>=13.7.0",
    "httpx>=0.26.0",
    "pyyaml>=6.0",
]

[project.scripts]
cmdai-terminal = "cmdai_terminal.__main__:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
